Refactor and pull pages to own file
