+++ flutter_app/pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   file_picker: ^5.3.4
   screen_capturer: ^0.1.7
   sidebarx: ^0.16.1
+  path_provider: ^2.1.0
 
 dev_dependencies:
   flutter_test:
